Split bric.h into smaller files in src/*
